+++ src/main/kotlin/minesweeper/application/MineSweeper.kt
@@ -2,7 +2,7 @@ package minesweeper.application
 
 import minesweeper.application.dto.MineFieldView
 import minesweeper.application.dto.Row
-import minesweeper.domain.MineFieldFactory
+import minesweeper.domain.MineField
 import minesweeper.domain.RandomMineCoordinateGenerator
 import minesweeper.domain.vo.Height
 import minesweeper.domain.vo.NumberOfMine
@@ -11,7 +11,7 @@ import minesweeper.domain.vo.Width
 class MineSweeper {
 
     fun createMineField(height: Height, width: Width, numberOfMine: NumberOfMine): MineFieldView {
-        val mineField = MineFieldFactory.create(
+        val mineField = MineField.create(
             height, width, numberOfMine,
             RandomMineCoordinateGenerator
         )
